streamlit_rag_chat/app/requirements.txt
streamlit 
openai 
PyPDF2 
python-docx 
sentence-transformers 
faiss-cpu
numpy
sklearn